requirements.txt
pandas
scikit-learn
mysql-connector-python
flask
gunicorn